previewを実装下に配置
